inserito funzioni di statistiche, aggiornato oggetto repos con stats anziche issues, cominciato a scrivere chiamata a repoApi per scrittura su mongo (da finire e rivedere)
