+++ dashboard_hack/githubApi/config.json.example
@@ -1,0 +1,4 @@
+{
+	"userToken": "your_user_token",
+	"ownerName": "repository_owner_name"
+}
