+++ dashboard_hack/public/index.html
@@ -22,12 +22,39 @@
       			<span id="loadingText"> Loading...</span>
       		</div>
 
-			<header>
-				<h1 class="headboard">Github dashboard issues</h1>
+			<header class="headboard">
+				<h1>Github dashboard issues</h1>
+				<nav class="myNavbar">
+					<ul class="nav justify-content-end">
+						<li class="nav-item">
+								<a class="nav-link active" href="#">back</a>
+						</li>
+						<li class="nav-item dropdown">
+							<a class="nav-link dropdown-toggle" data-toggle="dropdown" href="#" role="button" aria-haspopup="true" aria-expanded="false">repository</a>
+							<div class="dropdown-menu">
+								<div class="dropline">
+									<form class="col-lg-12 col-md-12 col-sm-12">
+   										<div class="form-group">
+      										<input type="search" class="form-control" id="exampleDropdownFormEmail1" 	placeholder="repository name">
+    									</div>
+    								<button type="submit" class="btn btn-primary"><img src="img/search.png"></button>
+  									</form>
+  								</div>
+							    <a class="dropdown-item" href="#">repository a</a>
+							    <a class="dropdown-item" href="#">repository a</a>
+							    
+							    <a class="dropdown-item" href="#">repository a</a>
+							</div>
+						</li>
+						<li class="nav-item">
+								<a class="nav-link" href="#">logout</a>
+						</li>
+					</ul>
+				</nav>
 			</header>
 			<div class="container-fluid">
 				<div class="row">
-					<div class="col-lg-2 col-md-2 col-sm-2 d-none d-sm-block sidebar myNavbar">
+					<!--<div class="col-lg-2 col-md-2 col-sm-2 d-none d-sm-block sidebar myNavbar">
 						<nav>
 							<ul class="nav flex-column">
 								<li class="nav-item">
@@ -44,16 +71,12 @@
 								</li>
 							</ul>
 						</nav>
-					</div>	
+					</div>	-->
 					<div class="total">
-						<div class="col-md-2 ml-md-auto col-lg-10 col-md-10 col-sm-10 myBodyDash">
+						<div class="col-md-2 ml-md-auto col-lg-12 col-md-12 col-sm-12 myBodyDash">
 							<div class="titleRepository">
 								<h2 id="name">ANPR </h2>
-								<h5>(</h5>
-								<h5 id="nTicket"></h5>
-								<span>Ticket totali</span>
-								<h5>)</5>
-
+								<span class="badge badge-pill badge-primary">123</span>
 							</div>
 							<div class="col-lg-8 col-md-12 col-sm-12 eqMargin">
 								<div class="block">
@@ -129,6 +152,6 @@
 			<!--js for chart.js-->
 			<!--<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.4.0/Chart.min.js"></script>-->
 			<!--personal js-->
-			<script src="js/dashboard.js"></script>
+			<!--script src="js/dashboard.js"></script-->
 		</body>
 	</html>
